requirements.txt
mkdocs>=1.4.0
python-frontmatter~=1.0.0
